Create a devcontainer
The devcontainer contains all of the dependencies necessary to do a
build and run tests, and uses docker images that have arm versions to
be fast on M1 macs.

Fix a build break due to stdlib differences on Mac vs. libc++13, and
change the default config to build with clang, since it is only the
default on macos.